Move LineMatch to internal/search/results
