+++ cmd/frontend/graphqlbackend/search_results_stats_languages_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
 	"github.com/sourcegraph/sourcegraph/internal/rcache"
+	"github.com/sourcegraph/sourcegraph/internal/search/results"
 	"github.com/sourcegraph/sourcegraph/internal/types"
 	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 )
@@ -59,9 +60,9 @@ func TestSearchResultsStatsLanguages(t *testing.T) {
 	defer git.ResetMocks()
 
 	fileMatch := func(path string, lineNumbers ...int32) *FileMatchResolver {
-		var lines []*LineMatch
+		var lines []*results.LineMatch
 		for _, n := range lineNumbers {
-			lines = append(lines, &LineMatch{LineNumber: n})
+			lines = append(lines, &results.LineMatch{LineNumber: n})
 		}
 		return mkFileMatchResolver(db, FileMatch{
 			Path:        path,
